chore: use real event carfiles for tests
