+++ runner/src/scenario/mod.rs
@@ -1,9 +1,10 @@
-use crate::scenario::ceramic::util::goose_error;
+use crate::scenario::util::goose_error;
 use goose::GooseError;
 
 pub mod ceramic;
 pub mod ipfs_block_fetch;
 pub mod recon_sync;
+pub mod util;
 
 static FIRST_USER: std::sync::atomic::AtomicBool = std::sync::atomic::AtomicBool::new(true);
